move angular test files under test folder
