+++ angular.json
@@ -16,9 +16,7 @@
             "outputPath": "dist/bhagavad-gita",
             "index": "src/index.html",
             "browser": "src/main.ts",
-            "polyfills": [
-              "zone.js"
-            ],
+            "polyfills": ["zone.js"],
             "tsConfig": "src/tsconfig.app.json",
             "assets": [
               {
@@ -75,7 +73,10 @@
         "test": {
           "builder": "@angular-builders/jest:run",
           "options": {
-            "tsConfig": "tsconfig.spec.json"
+            "main": "src/main.ts",
+            "polyfills": "src/polyfills.ts",
+            "tsConfig": "tsconfig.spec.json",
+            "testMatch": ["tests/**/*.spec.ts"]
           }
         },
         "lint": {
