+++ tests/app/grouped-sloka/grouped-sloka.component.spec.ts
@@ -1,36 +1,35 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { SingleSlokaComponent } from './single-sloka.component';
-import { SlokaComponent } from '../sloka/sloka.component';
+import { GroupedSlokaComponent } from '../../../src/app/grouped-sloka/grouped-sloka.component';
+import { SlokaComponent } from '../../../src/app/sloka/sloka.component';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { UtilityService } from '../services/utility.service';
 
-describe('SingleSlokaComponent', () => {
-  let component: SingleSlokaComponent;
-  let fixture: ComponentFixture<SingleSlokaComponent>;
+describe('GroupedSlokaComponent', () => {
+  let component: GroupedSlokaComponent;
+  let fixture: ComponentFixture<GroupedSlokaComponent>;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
       imports: [
         CommonModule,
         FormsModule,
-        HttpClientTestingModule,
         SlokaComponent,
+        GroupedSlokaComponent,
+        HttpClientTestingModule,
       ],
-      providers: [UtilityService],
     }).compileComponents();
 
-    fixture = TestBed.createComponent(SingleSlokaComponent);
+    fixture = TestBed.createComponent(GroupedSlokaComponent);
     component = fixture.componentInstance;
     // Provide default inputs
-    component.index = 0;
-    component.sloka = 'Sloka 1 text';
-    component.expandedSloka = 0;
+    component.group = [0, 1];
+    component.slokas = { 0: 'Sloka 1 text', 1: 'Sloka 2 text' };
     component.chapterId = 1;
     component.showSanskrit = false;
     component.showSandhi = false;
     component.isSlokaGroupsReady = true;
+    component.expandedSloka = 0;
     fixture.detectChanges();
   });
 
@@ -38,16 +37,14 @@ describe('SingleSlokaComponent', () => {
     expect(component).toBeTruthy();
   });
 
-  it('should emit slokaToggle event when onToggle is called', () => {
-    jest.spyOn(component.slokaToggle, 'emit');
-    component.onToggle();
-    expect(component.slokaToggle.emit).toHaveBeenCalledWith(0);
+  it('should emit toggleGroup event when onToggleGroup is called', () => {
+    jest.spyOn(component.toggleGroup, 'emit');
+    component.onToggleGroup(1);
+    expect(component.toggleGroup.emit).toHaveBeenCalledWith(1);
   });
 
   it('should return correct sloka title', () => {
-    jest
-      .spyOn(component['utilityService'], 'getSlokaTitle')
-      .mockReturnValue('Sloka 1');
     expect(component.getSlokaTitle(0)).toBe('Sloka 1');
+    expect(component.getSlokaTitle(2)).toBe('Sloka 3');
   });
 });
