+++ tests/app/services/chapter.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
-import { ChapterService } from './chapter.service';
-import { environment } from '../../environments/environment';
+import { ChapterService } from '../../../src/app/services/chapter.service';
+import { environment } from '../../../src/environments/environment';
 
 describe('ChapterService', () => {
   let service: ChapterService;
