+++ tsconfig.spec.json
@@ -5,5 +5,5 @@
     "types": ["jest", "node"]
   },
   "files": ["setup-jest.ts"],
-  "include": ["src/**/*.spec.ts", "src/**/*.d.ts"]
+  "include": ["src/**/*.d.ts", "tests/**/*.spec.ts"]
 }
